[BTAT-1823] Refactored audit mocks to mocks package
